Task 3rd AI CHATBOT WITH NLP
In this task, an AI-powered chatbot was developed using Natural Language Processing (NLP) techniques. The chatbot was trained on a JSON-based dataset of intents and responses. It uses NLTK for tokenization and lemmatization, and a machine learning model (e.g., Naive Bayes) for classifying user input and generating appropriate responses. This project demonstrated how NLP can be applied to understand and respond to user queries, simulating real human-like conversations and showcasing the fundamentals of AI in interactive applications.